rename package name to lower case
