+++ app/src/main/java/org/nervos/neuron/activity/AppWebActivity.java
@@ -31,14 +31,13 @@ import org.jetbrains.annotations.NotNull;
 import org.nervos.neuron.R;
 import org.nervos.neuron.item.AppItem;
 import org.nervos.neuron.item.ChainItem;
-import org.nervos.neuron.item.NeuronDApp.BaseNeuronDAppCallbackItem;
-import org.nervos.neuron.item.NeuronDApp.QrCodeItem;
+import org.nervos.neuron.item.dapp.BaseNeuronDAppCallbackItem;
+import org.nervos.neuron.item.dapp.QrCodeItem;
 import org.nervos.neuron.item.TitleItem;
 import org.nervos.neuron.item.TokenItem;
 import org.nervos.neuron.item.WalletItem;
 import org.nervos.neuron.plugin.NeuronDAppPlugin;
 import org.nervos.neuron.service.http.HttpEtherUrls;
-import org.nervos.neuron.service.http.HttpUrls;
 import org.nervos.neuron.service.http.NeuronSubscriber;
 import org.nervos.neuron.service.http.SignService;
 import org.nervos.neuron.service.http.WalletService;
